FIXed dll linkage CHG: some pretty printing
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/VistaAddonLibs/VistaMessagingService/trunk@23486 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VmsMsgReceptor.cpp
@@ -5,7 +5,7 @@
 /*                                             .                              */
 /*                                               RRRR WW  WW   WTTTTTTHH  HH  */
 /*                                               RR RR WW WWW  W  TT  HH  HH  */
-/*      Header   :	VmsMsgFactory.cpp RRRR   WWWWWWWW  TT  HHHHHH  */
+/*      Header   :	VmsMsgReceptor.cpp			 RRRR   WWWWWWWW  TT  HHHHHH  */
 /*                                               RR RR   WWW WWW  TT  HH  HH  */
 /*      Module   :  			                 RR  R    WW  WW  TT  HH  HH  */
 /*                                                                            */
@@ -87,11 +87,11 @@ int VmsMsgReceptor::ProcessIncomingMsg()
 	
 int VmsMsgReceptor::RegisterHandler(VmsMsgHandler *pHandler, int id)
 {
-	int len = m_oHandlers.size();
+	size_t len = m_oHandlers.size();
 	if(id>=len)
 	{
 		m_oHandlers.resize(id+1);
-		for(int i=len;i<id;i++)
+		for(size_t i=len;i<id;i++)
 		{
 			m_oHandlers[i]=NULL;
 		}
